+++ app/src/js/actions/index.js
@@ -291,14 +291,15 @@ export const applyRecoveryWorkflowToGranule = (granuleId) => (dispatch) => dispa
     error
   }));
 
-export const reingestGranule = (granuleId) => ({
+export const reingestGranule = (granuleId, meta) => ({
   [CALL_API]: {
     type: types.GRANULE_REINGEST,
     method: 'PUT',
     id: granuleId,
     path: `granules/${granuleId}`,
     data: {
-      action: 'reingest'
+      action: 'reingest',
+      ...meta,
     }
   }
 });
@@ -742,6 +743,7 @@ export const getExecutionStatus = (arn) => ({
   [CALL_API]: {
     type: types.EXECUTION_STATUS,
     method: 'GET',
+    id: arn,
     url: new URL(`executions/status/${arn}`, root).href
   }
 });
@@ -766,10 +768,11 @@ export const listExecutions = (options) => (dispatch, getState) => {
   });
 };
 
-export const listExecutionsByGranule = (payload) => ({
+export const listExecutionsByGranule = (granuleId, payload) => ({
   [CALL_API]: {
     type: types.EXECUTIONS_LIST,
     method: 'POST',
+    id: granuleId,
     path: 'executions/search-by-granules',
     params: { limit: defaultPageLimit },
     data: payload
@@ -780,6 +783,20 @@ export const filterExecutions = (param) => ({ type: types.FILTER_EXECUTIONS, par
 export const clearExecutionsFilter = (paramKey) => ({ type: types.CLEAR_EXECUTIONS_FILTER, paramKey });
 export const searchExecutions = (infix) => ({ type: types.SEARCH_EXECUTIONS, infix });
 export const clearExecutionsSearch = () => ({ type: types.CLEAR_EXECUTIONS_SEARCH });
+
+export const getGranulesWorkflows = (payload) => ({
+  [CALL_API]: {
+    type: types.WORKFLOWS_FROM_GRANULES,
+    method: 'POST',
+    path: 'executions/workflows-by-granules',
+    params: { limit: 50 },
+    data: payload
+  }
+});
+
+export const clearGranulesWorkflows = () => ({ type: types.CLEAR_WORKFLOWS_FROM_GRANULES });
+
+export const getGranulesWorkflowsClearError = () => ({ type: types.WORKFLOWS_FROM_GRANULES_CLEAR_ERROR });
 
 export const listOperations = (options) => (dispatch, getState) => {
   const timeFilters = fetchCurrentTimeFilters(getState().datepicker);