+++ app/src/js/components/Granules/bulk-granule-modal.js
@@ -1,9 +1,18 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { get } from 'object-path';
 
 import _config from '../../config';
 import DefaultModal from '../Modal/modal';
+import ErrorReport from '../Errors/report';
+import SimpleDropdown from '../DropDown/simple-dropdown';
 import TextArea from '../TextAreaForm/text-area';
+import {
+  clearGranulesWorkflows,
+  getGranulesWorkflows,
+  getGranulesWorkflowsClearError
+} from '../../actions';
 
 const { kibanaRoot } = _config;
 
@@ -18,17 +27,21 @@ const BulkGranuleModal = ({
   defaultQuery,
   dispatch,
   error,
+  granulesExecutions,
   handleSuccessConfirm,
   inflight,
   onCancel,
+  queryWorkflowOptions = false,
   requestId,
   showModal,
   selected = [],
+  selectWorkflow = false,
   success,
   successMessage,
-  title
+  title,
 }) => {
   const [query, setQuery] = useState(JSON.stringify(defaultQuery, null, 2));
+  const [workflow, setWorkflow] = useState(null);
   const [errorState, setErrorState] = useState();
   let buttonText;
 
@@ -48,6 +61,8 @@ const BulkGranuleModal = ({
     if (!inflight) {
       try {
         json = JSON.parse(query);
+        const granuleIds = json.ids.map((granule) => granule.granuleId);
+        json.ids = granuleIds;
       } catch (jsonError) {
         return setErrorState('Syntax error in JSON');
       }
@@ -55,17 +70,52 @@ const BulkGranuleModal = ({
     }
   }
 
-  function onChange (id, value) {
-    if (errorState) {
-      try {
-        JSON.parse(value);
-        setErrorState();
-      } catch (_) {
-        // empty
-      }
+  function queryGranulesWorkflows(queryParams) {
+    const { ids, index, query: esQuery } = queryParams;
+    if ((index && esQuery) || ids.length > 0) {
+      const granuleWorkflowsQuery = { ...queryParams, granules: ids };
+      delete granuleWorkflowsQuery[ids];
+      dispatch(getGranulesWorkflows(granuleWorkflowsQuery));
     }
-    setQuery(value);
   }
+
+  function onChange (id, value) {
+    setQuery(value);
+    try {
+      const queryParams = JSON.parse(value);
+      if (errorState) {
+        setErrorState();
+      }
+      if (queryWorkflowOptions) {
+        queryGranulesWorkflows(queryParams);
+      }
+    } catch (_) {
+      // empty
+    }
+  }
+
+  function handleSelectWorkflow(selector, selectedWorkflow) {
+    setWorkflow(selectedWorkflow);
+    setQuery(JSON.stringify({ workflowName: selectedWorkflow, ...JSON.parse(query) }, null, 2));
+  }
+
+  useEffect(() => {
+    const queryParams = { ...JSON.parse(query), ids: selected };
+    setQuery(JSON.stringify(queryParams, null, 2));
+
+    if (showModal && queryWorkflowOptions) {
+      queryGranulesWorkflows(queryParams);
+    }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [dispatch, showModal, queryWorkflowOptions, JSON.stringify(selected)]);
+
+  useEffect(() => {
+    if (!showModal && queryWorkflowOptions) {
+      dispatch(clearGranulesWorkflows());
+      dispatch(getGranulesWorkflowsClearError());
+      setWorkflow(null);
+    }
+  }, [dispatch, showModal, queryWorkflowOptions]);
 
   return (
     <DefaultModal
@@ -108,7 +158,7 @@ const BulkGranuleModal = ({
           {selected &&
             <>
               <p>Selected granules:</p>
-              <p>[{selected.map((selection) => `"${selection}"`).join(', ')}]</p>
+              <p>[{selected.map((selection) => `"${selection.granuleId}"`).join(', ')}]</p>
             </>
           }
           <br/>
@@ -123,6 +173,25 @@ const BulkGranuleModal = ({
               maxLines={200}
             />}
           </form>
+          {selectWorkflow &&
+          <>
+            <h4 className="modal_subtitle">Then select workflow to rerun for all the selected granules.</h4>
+            {get(granulesExecutions, 'workflows.error') &&
+              <ErrorReport report={`Failed to get workflows: ${get(granulesExecutions, 'workflows.error')}`}/>}
+            <div className='modal__internal modal__formcenter'>
+              <SimpleDropdown
+                isClearable={true}
+                key={'workflow-dropdown'}
+                label={'Select Workflow'}
+                value={workflow}
+                options={granulesExecutions.workflows.data}
+                id='workflow-dropdown'
+                onChange={handleSelectWorkflow}
+                placeholder="Workflow Name"
+              />
+            </div>
+          </>
+          }
         </div>
       }
     </DefaultModal>
@@ -140,15 +209,27 @@ BulkGranuleModal.propTypes = {
   defaultQuery: PropTypes.object,
   dispatch: PropTypes.func,
   error: PropTypes.string,
+  granulesExecutions: PropTypes.object,
   handleSuccessConfirm: PropTypes.func,
   inflight: PropTypes.bool,
   onCancel: PropTypes.func,
+  // whether query workflow options for the selected granule
+  queryWorkflowOptions: PropTypes.bool,
   requestId: PropTypes.string,
-  selected: PropTypes.array,
+  selected: PropTypes.arrayOf(PropTypes.shape({
+    granuleId: PropTypes.string,
+    collectionId: PropTypes.string,
+  })),
+  // whether select a workflow from dropdown
+  selectWorkflow: PropTypes.bool,
   showModal: PropTypes.bool,
   success: PropTypes.bool,
   successMessage: PropTypes.string,
   title: PropTypes.string
 };
 
-export default BulkGranuleModal;
+export { BulkGranuleModal };
+
+export default connect((state) => ({
+  granulesExecutions: state.granulesExecutions,
+}))(BulkGranuleModal);