+++ cypress/integration/granules_spec.js
@@ -376,6 +376,28 @@ describe('Dashboard Granules Page', () => {
       cy.get('.table .tbody .tr').should('have.length', 2);
     });
 
+    it('Should reingest a granule from granule details page.', () => {
+      const granuleId = 'MOD09GQ.A9344328.K9yI3O.006.4625818663028';
+      cy.intercept(
+        { method: 'PUT', url: new RegExp('/granules/.*') },
+        { body: { message: 'ingested' }, statusCode: 200 },
+      );
+      cy.intercept('GET', `/granules/${granuleId}*`).as('getGranule');
+
+      cy.visit(`/granules/granule/${granuleId}`);
+      cy.wait('@getGranule');
+      cy.get('.heading--large').should('have.text', `Granule: ${granuleId}`);
+      cy.contains('button', 'Options').click();
+      cy.get('.dropdown__menu').contains('Reingest').click();
+      cy.get('.modal-body .form__dropdown .dropdown__element input').as('workflow-input');
+      cy.get('@workflow-input').click({ force: true }).type('IngestAndPublish').type('{enter}');
+      cy.get('.modal-body .form__dropdown .dropdown__element').should('have.text', 'IngestAndPublishGranule');
+      cy.get('.button--submit').click();
+      cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Success');
+      cy.get('.button--cancel').click();
+      cy.url().should('include', `granules/granule/${granuleId}`);
+    });
+
     it('Should reingest a granule and redirect to the granules detail page.', () => {
       const granuleId = 'MOD09GQ.A0142558.ee5lpE.006.5112577830916';
       cy.intercept(
@@ -387,6 +409,9 @@ describe('Dashboard Granules Page', () => {
       cy.get(`[data-value="${granuleId}"] > .td >input[type="checkbox"]`).click();
       cy.contains('button', 'Granule Actions').click();
       cy.contains('button', 'Reingest').click();
+      cy.get('.modal-body .form__dropdown .dropdown__element input').as('workflow-input');
+      cy.get('@workflow-input').click({ force: true }).type('Discover').type('{enter}');
+      cy.get('.modal-body .form__dropdown .dropdown__element').should('have.text', 'DiscoverAndQueuePdrs');
       cy.get('.button--submit').click();
       cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Success');
       cy.get('.button__goto').click();
@@ -397,7 +422,7 @@ describe('Dashboard Granules Page', () => {
 
     it('Should reingest multiple granules and redirect to the running page.', () => {
       const granuleIds = [
-        'MOD09GQ.A0142558.ee5lpE.006.5112577830916',
+        'MOD09GQ.A2417309.YZ9tCV.006.4640974889044',
         'MOD09GQ.A9344328.K9yI3O.006.4625818663028'
       ];
       cy.intercept(
@@ -409,11 +434,43 @@ describe('Dashboard Granules Page', () => {
       cy.get(`[data-value="${granuleIds[1]}"] > .td >input[type="checkbox"]`).click();
       cy.contains('button', 'Granule Actions').click();
       cy.contains('button', 'Reingest').click();
+      cy.get('.modal-body .form__dropdown .dropdown__element input').as('workflow-input');
+      cy.get('@workflow-input').click({ force: true }).type('IngestAndP').type('{enter}');
+      cy.get('.modal-body .form__dropdown .dropdown__element').should('have.text', 'IngestAndPublishGranule');
       cy.get('.button--submit').click();
       cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Success');
       cy.get('.button__goto').click();
       cy.url().should('include', 'granules/processing');
       cy.get('.heading--large').should('have.text', 'Running Granules 2');
+    });
+
+    it('Should reingest multiple granules selected from multiple pages.', () => {
+      cy.intercept(
+        { method: 'PUT', url: new RegExp('/granules/.*') },
+        { body: { message: 'ingested' }, statusCode: 200 },
+      );
+      cy.visit('/granules?limit=2');
+      cy.get('.table .tbody .tr').should('have.length', 2);
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.get('.pagination ol li').last().should('have.text', 'Next').click();
+      cy.wait(1000);
+      cy.url().should('include', 'page=2');
+      cy.get('.table .tbody .tr').should('have.length', 2);
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox2');
+      cy.get('@granule-checkbox2').click({ multiple: true });
+
+      cy.get('.pagination ol li').last().should('have.text', 'Next').click();
+      cy.wait(1000);
+      cy.url().should('include', 'page=3');
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Reingest').click();
+      cy.get('.modal-content .modal-body').should('contain.text', 'Selected granules:');
+      cy.get('.modal-content .modal-body > ul > li').should('have.length', 4);
+      cy.get('.button--submit').click();
+      cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Success');
     });
 
     it('Should fail to reingest multiple granules and remain on the page.', () => {
@@ -681,7 +738,7 @@ describe('Dashboard Granules Page', () => {
       cy.contains('button', 'Delete').click();
       cy.contains('.button--submit', 'Confirm').click();
 
-      cy.get('.default-modal.batch-async-modal ').as('modal');
+      cy.get('.default-modal.batch-async-modal ', { timeout: 10000 }).as('modal');
       cy.get('@modal').contains('div', 'Success!');
     });